+++ src/cryptography/hazmat/primitives/asymmetric/ec.py
@@ -5,7 +5,6 @@
 from __future__ import absolute_import, division, print_function
 
 import abc
-import warnings
 
 import six
 
@@ -357,47 +356,6 @@ class EllipticCurvePublicNumbers(object):
 
     def public_key(self, backend):
         return backend.load_elliptic_curve_public_numbers(self)
-
-    def encode_point(self):
-        warnings.warn(
-            "encode_point has been deprecated on EllipticCurvePublicNumbers"
-            " and will be removed in a future version. Please use "
-            "EllipticCurvePublicKey.public_bytes to obtain both "
-            "compressed and uncompressed point encoding.",
-            utils.DeprecatedIn25,
-            stacklevel=2,
-        )
-        # key_size is in bits. Convert to bytes and round up
-        byte_length = (self.curve.key_size + 7) // 8
-        return (
-            b'\x04' + utils.int_to_bytes(self.x, byte_length) +
-            utils.int_to_bytes(self.y, byte_length)
-        )
-
-    @classmethod
-    def from_encoded_point(cls, curve, data):
-        if not isinstance(curve, EllipticCurve):
-            raise TypeError("curve must be an EllipticCurve instance")
-
-        warnings.warn(
-            "Support for unsafe construction of public numbers from "
-            "encoded data will be removed in a future version. "
-            "Please use EllipticCurvePublicKey.from_encoded_point",
-            utils.DeprecatedIn25,
-            stacklevel=2,
-        )
-
-        if data.startswith(b'\x04'):
-            # key_size is in bits. Convert to bytes and round up
-            byte_length = (curve.key_size + 7) // 8
-            if len(data) == 2 * byte_length + 1:
-                x = utils.int_from_bytes(data[1:byte_length + 1], 'big')
-                y = utils.int_from_bytes(data[byte_length + 1:], 'big')
-                return cls(x, y, curve)
-            else:
-                raise ValueError('Invalid elliptic curve point data length')
-        else:
-            raise ValueError('Unsupported elliptic curve point type')
 
     curve = utils.read_only_property("_curve")
     x = utils.read_only_property("_x")